Using fixtures to use the same tests across multiple client configurations.
